requirements.txt
numpy==1.25.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
joblib==1.3.2
streamlit==1.30.0
scapy==2.5.0
tensorflow==2.14.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
passlib==1.7.4
python-dotenv==1.0.0
